# schemas/s3-api.openapi.yaml
openapi: 3.1.0
info:
  title: BleepStore S3-Compatible API
  version: 1.0.0
  description: |
    BleepStore implements a core subset of the Amazon S3 REST API, providing
    S3-compatible object storage via path-style URLs. This schema covers
    Phase 1 operations including bucket management, object CRUD, access-control
    lists, and multipart uploads.

    All XML request and response bodies use the S3 namespace:
    http://s3.amazonaws.com/doc/2006-03-01/
  license:
    name: MIT
servers:
  - url: https://localhost:9000
    description: Local BleepStore development server
  - url: https://s3.bleepstore.example.com
    description: BleepStore production endpoint
security:
  - sigv4: []

# ---------------------------------------------------------------------------
# Tags
# ---------------------------------------------------------------------------
...
paths:
...
  /:
    get:
      operationId: ListBuckets
      summary: List all buckets owned by the authenticated sender
      tags: [Bucket]
      responses:
        "200":
          description: Bucket list returned successfully.
          headers:
            x-amz-request-id:
              $ref: "#/components/headers/x-amz-request-id"
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/ListAllMyBucketsResult"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # =======================================================================
  # Bucket-level: /{Bucket}
  # =======================================================================
...
components:
...
  securitySchemes:
    sigv4:
      type: http
      scheme: bearer
      description: |
        AWS Signature Version 4 authentication. Clients must sign each
        request using their access key ID and secret access key, placing the
        signature in the Authorization header (or as query-string
        parameters for pre-signed URLs). The format is:

          Authorization: AWS4-HMAC-SHA256
            Credential=<access-key>/<date>/<region>/s3/aws4_request,
            SignedHeaders=<signed-headers>,
            Signature=<signature>

        BleepStore validates the signature against stored credentials.

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
...
  headers:
    x-amz-request-id:
      description: Unique identifier assigned to the request by BleepStore.
      schema:
        type: string

  # -------------------------------------------------------------------------
  # Responses
  # -------------------------------------------------------------------------
  responses:
...
    AccessDenied:
      description: "403 -- Access denied. The credentials are invalid or lack permission."
      headers:
        x-amz-request-id:
          $ref: "#/components/headers/x-amz-request-id"
      content:
        application/xml:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: "500 -- An internal server error occurred."
      headers:
        x-amz-request-id:
          $ref: "#/components/headers/x-amz-request-id"
      content:
        application/xml:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: "503 -- The service is temporarily unavailable. Retry with exponential backoff."
      headers:
        x-amz-request-id:
          $ref: "#/components/headers/x-amz-request-id"
      content:
        application/xml:
          schema:
            $ref: "#/components/schemas/Error"

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      xml:
        name: Error
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        Code:
          type: string
          description: Machine-readable error code (e.g. NoSuchBucket, AccessDenied).
        Message:
          type: string
          description: Human-readable error description.
        Resource:
          type: string
          description: The resource associated with the error.
        RequestId:
          type: string
          description: Unique request identifier.
        HostId:
          type: string
          description: Extended request ID for debugging.
        BucketName:
          type: string
          description: Bucket name (present on some bucket-related errors).
        Key:
          type: string
          description: Object key (present on some object-related errors).
      required: [Code, Message, RequestId]

    # === Bucket list ======================================================
    ListAllMyBucketsResult:
      type: object
      xml:
        name: ListAllMyBucketsResult
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        Owner:
          $ref: "#/components/schemas/Owner"
        Buckets:
          type: object
          properties:
            Bucket:
              type: array
              items:
                $ref: "#/components/schemas/BucketInfo"
              xml:
                wrapped: false
      required: [Owner, Buckets]
    BucketInfo:
      type: object
      xml:
        name: Bucket
      properties:
        Name:
          type: string
        CreationDate:
          type: string
          format: date-time
      required: [Name, CreationDate]

    # === Bucket configuration =============================================
...
    Owner:
      type: object
      properties:
        ID:
          type: string
          description: Canonical user ID.
        DisplayName:
          type: string
          description: Display name of the owner.
      required: [ID]
...